// web/apps/msteams/src/components/organization/organizationPage/__generated__/organizationZonesTab_zones_query.graphql.ts
export type organizationZonesTab_zones_query$data = {
  readonly zones: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"zoneCard_OrganizationTagDetails">;
      };
    }>;
    readonly totalCount: number | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "organizationZonesTab_zones_query";
};
export type organizationZonesTab_zones_query$key = {
  readonly " $data"?: organizationZonesTab_zones_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"organizationZonesTab_zones_query">;
};